KDSH-TrackA/requirements.txt
pathway[xpack-llm]
google-genai
python-dotenv
rich
